fix: 31 correctness + overflow + crash bugs; add gpu_config wiring, build speed, EXPLAIN, async footer, per-query memory budget
Bugs fixed:
- AVG aggregate: count denominator used num_rows() (incl nulls) instead of non-null count
- top_n: ro_views vector never populated (latent UAF on concat)
- prefetching_cache: cudaMemcpyAsync failure only fprintf'd (silent data corruption); error path didn't sync in-flight memcpy before release_read (UAF)
- uring_reactor: shutdown credited failed H2D copy as complete (silent corruption)
- rest_reactor: signed overflow in Retry-After backoff; piece-count overflow; content-range parser overflow; exponential backoff overflow
- dict_fsst/fsst: uint32_t accumulators (total_dict_entries, total_fsst_row_count) overflow on large tables → undersized buffers + OOB writes
- gpu_decode_strings: static_cast<uint32_t> truncation of cross-run offset bases (wrong string data)
- request_manager: data race on non-atomic first_exception in destructor (UB)
- hash_join/concat: null src_pipeline deref in task hints (crash)
- concat: null deref after dynamic_cast (crash)
- sort_sample: partition count overflow (OOM / wrong partitioning)
- hash_join + nested_loop_join: num_batches_to_process overflow (dropped join rows)
- small_in_list: static_assert documenting k_max_keys invariant
- io_utils: overflow-safe end computation in coalesce_sorted/common_sorted/missing_sorted

Optimizations added:
- gpu_config singleton wired into live paths (init_adaptive, get_target_ctas, determine_num_partitions)
- ccache + unity build in CMakeLists.txt + rocm-dev preset
- EXPLAIN falls back to DuckDB native instead of throwing
- Async Parquet footer read overlapped with column-projection setup
- Per-query GPU memory budget (SET sirius.per_query_memory_budget)

+++ src/cuda/scan/strings/dict_fsst.cu
@@ -556,7 +556,10 @@ prepared_dict_fsst prepare_dict_fsst(gpu_string_codec_run const& run,
   // Phase 2: validate, compute per-segment region offsets + cumulative
   // base_off into the global byte/decoded_offsets arrays.
   std::vector<dict_fsst_pre_desc> pre(num_segs);
-  uint32_t total_dict_entries = 0;  // sum of (dict_count + 1) for valid segs
+  // Use size_t (not uint32_t) for the cross-segment accumulator: with many
+  // dictionary segments, the uint32_t sum of (dict_count + 1) wraps past
+  // UINT32_MAX, producing undersized device/host buffers and OOB writes.
+  size_t total_dict_entries = 0;
   for (uint32_t i = 0; i < num_segs; ++i) {
     auto const& seg = run.segments[i];
     auto& p         = pre[i];
@@ -582,7 +585,7 @@ prepared_dict_fsst prepare_dict_fsst(gpu_string_codec_run const& run,
     p.dict_count           = hdr.dict_count;
     p.mode                 = hdr.mode;
     p.string_lengths_width = hdr.string_lengths_width;
-    p.base_off             = total_dict_entries;
+    p.base_off             = static_cast<uint32_t>(total_dict_entries);
     p.valid                = 1;
     total_dict_entries += hdr.dict_count + 1u;
   }
